:recycle:[refactor]: 커뮤니티 작성 페이지 너비 조정
