+++ src/app/(route)/board/post/employment/page.tsx
@@ -1,5 +1,5 @@
 'use client'
-import PostEmploymentContainer from '@/app/components/board/containers/PostContainer'
+import PostEmploymentContainer from '@/app/components/board/containers/PostEmploymentContainer'
 import Footer from '@/app/components/layout/Footer'
 import Header from '@/app/components/layout/Header'
 import { useSearchParams } from 'next/navigation'
@@ -10,7 +10,7 @@ export default function PostEmploymentPage() {
   return (
     <section className="flex flex-col items-center min-h-screen relative">
       <Header nickname={nickname || ''} />
-      <div className="w-[94%] h-full py-9 mb-[200px]  flex items-center justify-center">
+      <div className="w-[88%] h-full py-10 mb-[200px] flex items-center justify-center">
         <PostEmploymentContainer />
       </div>
       <Footer />
