+++ src/app/(route)/board/post/group/page.tsx
@@ -10,7 +10,7 @@ export default function PostGroupPage() {
   return (
     <section className="flex flex-col items-center min-h-screen relative">
       <Header nickname={nickname || ''} />
-      <div className="w-[94%] h-full py-9 mb-[200px]  flex items-center justify-center">
+      <div className="w-[88%] h-full py-10 mb-[200px] flex items-center justify-center">
         <PostGroupContainer />
       </div>
       <Footer />
