+++ src/app/(route)/board/update/employment/page.tsx
@@ -1,5 +1,5 @@
 'use client'
-import PostEmploymentContainer from '@/app/components/board/containers/PostContainer'
+import PostEmploymentContainer from '@/app/components/board/containers/PostEmploymentContainer'
 import Footer from '@/app/components/layout/Footer'
 import Header from '@/app/components/layout/Header'
 import { useSearchParams } from 'next/navigation'
@@ -31,7 +31,7 @@ export default function PatchEmploymentPage() {
     return (
       <section className="flex flex-col items-center min-h-screen relative">
         <Header nickname={data?.memberName || ''} />
-        <div className="w-[94%] h-full py-9 mb-[200px]  flex items-center justify-center">
+        <div className="w-[88%] h-full py-10 mb-[200px] flex items-center justify-center">
           <PostEmploymentContainer isEdit={true} data={data} />
         </div>
         <Footer />
