+++ src/app/components/board/containers/PostEmploymentContainer.tsx
@@ -57,7 +57,7 @@ const PostEmploymentContainer = ({
   }
 
   return (
-    <div className="flex flex-col w-full bg-white h-[786px] rounded-3xl px-[58px] py-9">
+    <div className="flex flex-col w-full bg-white h-[786px] rounded-3xl px-[58px] py-9 text-[16px]">
       <div className="text-xl font-bold">{'직무 카테고리  '}</div>
       <div className="flex gap-x-[33px] mt-[18px] mb-7">
         {JOBKEYWORD.map((str, index) => (
@@ -81,7 +81,7 @@ const PostEmploymentContainer = ({
       </div>
       <textarea
         placeholder="게시물의 내용을 입력해주세요"
-        className="w-full h-[440px] px-7 py-2 text-xl font-medium rounded-[6px] border-[2px] border-[#C4C4C4]"
+        className="w-full h-[400px] px-6 py-3 text-[18px] font-medium rounded-[6px] border-[2px] border-[#C4C4C4]"
         onChange={(e) => setContent(e.target.value)}
         value={content}
       />
